wip
bugfixes
AMP interview should now be ok
cleanups

+++ src/RootViewControllerPlaylist.mm
@@ -1,6 +1,6 @@
 //
-//  RootViewController.m
-//  modizer1
+//  RootViewController.mm
+//  modizer
 //
 //  Created by Yohann Magnien on 04/06/10.
 //  Copyright __YoyoFR / Yohann Magnien__ 2010. All rights reserved.
@@ -1147,32 +1147,52 @@ static int qsort_CompareArcEntries(const void *entryA, const void *entryB) {
     pthread_mutex_lock(&db_mutex);
     if (sqlite3_open([pathToDB UTF8String], &db) == SQLITE_OK){
         char sqlStatement[1024];
+        bool allOK=true;
         
         err=sqlite3_exec(db, "PRAGMA journal_mode=WAL; PRAGMA cache_size = 1;PRAGMA synchronous = 1;PRAGMA locking_mode = EXCLUSIVE;", 0, 0, 0);
         if (err==SQLITE_OK){
         } else MDZELog("ErrSQL : %d",err);
         
-        snprintf(sqlStatement,sizeof(sqlStatement),"DELETE FROM playlists_entries WHERE id_playlist=%s",
-                 [playlist->playlist_id UTF8String]);
-        err=sqlite3_exec(db, sqlStatement, NULL, NULL, NULL);
-        if (err==SQLITE_OK){
-        } else MDZELog("ErrSQL : %d",err);
-        
         err=sqlite3_exec(db, "BEGIN TRANSACTION", 0, 0, 0);
         if (err==SQLITE_OK){
-        } else MDZELog("ErrSQL : %d",err);
+        } else {
+            allOK=false;
+            MDZELog("ErrSQL : %d",err);
+        }
         
-        for (int i=0;i<playlist->nb_entries;i++) {
-            snprintf(sqlStatement,sizeof(sqlStatement),"INSERT INTO playlists_entries (id_playlist,name,fullpath) SELECT %s,\"%s\",\"%s\"",
-                     [playlist->playlist_id UTF8String],[[playlist->entries[i].label lastPathComponent] UTF8String],[playlist->entries[i].fullpath UTF8String]);
+        if (allOK) {
+            snprintf(sqlStatement,sizeof(sqlStatement),"DELETE FROM playlists_entries WHERE id_playlist=%s",
+                     [playlist->playlist_id UTF8String]);
             err=sqlite3_exec(db, sqlStatement, NULL, NULL, NULL);
+            if (err==SQLITE_OK){
+            } else {
+                allOK=false;
+                MDZELog("ErrSQL : %d",err);
+            }
+        }
+        
+        if (allOK) {
+            for (int i=0;i<playlist->nb_entries;i++) {
+                snprintf(sqlStatement,sizeof(sqlStatement),"INSERT INTO playlists_entries (id_playlist,name,fullpath) SELECT %s,\"%s\",\"%s\"",
+                         [playlist->playlist_id UTF8String],[[playlist->entries[i].label lastPathComponent] UTF8String],[playlist->entries[i].fullpath UTF8String]);
+                err=sqlite3_exec(db, sqlStatement, NULL, NULL, NULL);
+                if (err==SQLITE_OK){
+                } else {
+                    allOK=false;
+                    MDZELog("ErrSQL : %d",err);
+                }
+            }
+        }
+        
+        if (allOK) {
+            err=sqlite3_exec(db, "COMMIT", 0, 0, 0);
+            if (err==SQLITE_OK){
+            } else MDZELog("ErrSQL : %d",err);
+        } else {
+            err=sqlite3_exec(db, "ROLLBACK", 0, 0, 0);
             if (err==SQLITE_OK){
             } else MDZELog("ErrSQL : %d",err);
         }
-        
-        err=sqlite3_exec(db, "COMMIT", 0, 0, 0);
-        if (err==SQLITE_OK){
-        } else MDZELog("ErrSQL : %d",err);
         
         snprintf(sqlStatement,sizeof(sqlStatement),"UPDATE playlists SET num_files=\
     (SELECT COUNT(1) FROM playlists_entries e WHERE playlists.id=e.id_playlist AND playlists.id=%s)\
@@ -1715,6 +1735,7 @@ static int qsort_CompareArcEntries(const void *entryA, const void *entryB) {
                 extractProgress.cancellable = YES;
                 extractProgress.pausable = NO;
                 NSString *tmpPath=[NSString stringWithFormat:@"%@/tmpArchiveBrowser",NSTemporaryDirectory()];
+                [mFileMngr removeItemAtPath:tmpPath error:NULL];
                 [ModizFileHelper extractToPath:[cpath UTF8String] path:[tmpPath UTF8String] caller:self progress:extractProgress context:ExtractBrowserListProgressObserverContext];
             }
             
@@ -3626,7 +3647,8 @@ int getPlaylistStatsDBmod(t_playlist *pl) {
     const NSInteger BOTTOM_IMAGE_TAG = 1003;
     const NSInteger ACT_IMAGE_TAG = 1004;
     const NSInteger SECACT_IMAGE_TAG = 1005;
-    UILabel *topLabel;
+    //UILabel *topLabel;
+    CBAutoScrollLabel *topLabel;
     UILabel *bottomLabel;
     UIImageView *bottomImageView;
     UIButton *actionView,*secActionView;
@@ -3690,7 +3712,13 @@ int getPlaylistStatsDBmod(t_playlist *pl) {
         //
         // Create the label for the top row of text
         //
-        topLabel = [[UILabel alloc] init];
+        //topLabel = [[UILabel alloc] init];
+        topLabel = [[CBAutoScrollLabel alloc] init];
+        topLabel.labelSpacing = 35; // distance between start and end labels
+        topLabel.pauseInterval = 3.7; // seconds of pause before scrolling starts again
+        topLabel.scrollSpeed = 30; // pixels per second
+        topLabel.fadeLength = 12.f; // length of the left and right edge fade, 0 to disable
+        
         [cell.contentView addSubview:topLabel];
         //
         // Configure the properties for the text that are the same on every row
@@ -3698,8 +3726,8 @@ int getPlaylistStatsDBmod(t_playlist *pl) {
         topLabel.tag = TOP_LABEL_TAG;
         topLabel.backgroundColor = [UIColor clearColor];
         topLabel.font = [UIFont systemFontOfSize:17];
-        topLabel.lineBreakMode=(settings[GLOB_TruncateNameMode].detail.mdz_switch.switch_value?
-                                ((settings[GLOB_TruncateNameMode].detail.mdz_switch.switch_value==2) ? NSLineBreakByTruncatingTail:NSLineBreakByTruncatingMiddle):NSLineBreakByTruncatingHead);;;
+        //topLabel.lineBreakMode=(settings[GLOB_TruncateNameMode].detail.mdz_switch.switch_value?
+         //                       ((settings[GLOB_TruncateNameMode].detail.mdz_switch.switch_value==2) ? NSLineBreakByTruncatingTail:NSLineBreakByTruncatingMiddle):NSLineBreakByTruncatingHead);;;
         topLabel.opaque=TRUE;
         
         //
@@ -3738,14 +3766,15 @@ int getPlaylistStatsDBmod(t_playlist *pl) {
         cell.accessoryView=nil;
         //        cell.selectionStyle=UITableViewCellSelectionStyleGray;
     } else {
-        topLabel = (UILabel *)[cell viewWithTag:TOP_LABEL_TAG];
+        //topLabel = (UILabel *)[cell viewWithTag:TOP_LABEL_TAG];
+        topLabel = (CBAutoScrollLabel *)[cell viewWithTag:TOP_LABEL_TAG];
         bottomLabel = (UILabel *)[cell viewWithTag:BOTTOM_LABEL_TAG];
         bottomImageView = (UIImageView *)[cell viewWithTag:BOTTOM_IMAGE_TAG];
         actionView = (UIButton *)[cell viewWithTag:ACT_IMAGE_TAG];
         secActionView = (UIButton *)[cell viewWithTag:SECACT_IMAGE_TAG];
         
-        topLabel.lineBreakMode=(settings[GLOB_TruncateNameMode].detail.mdz_switch.switch_value?
-                                ((settings[GLOB_TruncateNameMode].detail.mdz_switch.switch_value==2) ? NSLineBreakByTruncatingTail:NSLineBreakByTruncatingMiddle):NSLineBreakByTruncatingHead);
+//        topLabel.lineBreakMode=(settings[GLOB_TruncateNameMode].detail.mdz_switch.switch_value?
+//                                ((settings[GLOB_TruncateNameMode].detail.mdz_switch.switch_value==2) ? NSLineBreakByTruncatingTail:NSLineBreakByTruncatingMiddle):NSLineBreakByTruncatingHead);
         
         bottomLabel.lineBreakMode=(settings[GLOB_TruncateNameMode].detail.mdz_switch.switch_value?
                                 ((settings[GLOB_TruncateNameMode].detail.mdz_switch.switch_value==2) ? NSLineBreakByTruncatingTail:NSLineBreakByTruncatingMiddle):NSLineBreakByTruncatingHead);
@@ -3755,12 +3784,12 @@ int getPlaylistStatsDBmod(t_playlist *pl) {
     
     if (darkMode) {
         topLabel.textColor = [UIColor colorWithRed:0.9 green:0.9 blue:0.9 alpha:1.0];
-        topLabel.highlightedTextColor = [UIColor colorWithRed:1 green:1 blue:1 alpha:1.0];
+        //topLabel.highlightedTextColor = [UIColor colorWithRed:1 green:1 blue:1 alpha:1.0];
         bottomLabel.textColor = [UIColor colorWithRed:0.6 green:0.6 blue:0.6 alpha:1.0];
         bottomLabel.highlightedTextColor = [UIColor colorWithRed:0.8 green:0.8 blue:0.8 alpha:1.0];
     } else {
         topLabel.textColor = [UIColor colorWithRed:0.1 green:0.1 blue:0.1 alpha:1.0];
-        topLabel.highlightedTextColor = [UIColor colorWithRed:0.0 green:0.0 blue:0.0 alpha:1.0];
+        //topLabel.highlightedTextColor = [UIColor colorWithRed:0.0 green:0.0 blue:0.0 alpha:1.0];
         bottomLabel.textColor = [UIColor colorWithRed:0.4 green:0.4 blue:0.4 alpha:1.0];
         bottomLabel.highlightedTextColor = [UIColor colorWithRed:0.2 green:0.2 blue:0.2 alpha:1.0];
     }
